backend/requirements.txt
# app
boto3==1.17.57
fastapi==0.55.1
uvicorn==0.11.3
pydantic==1.4
email-validator==1.1.1

# db
psycopg2-binary==2.8.6
SQLAlchemy==1.3.16
alembic==1.4.2

# dev
pytest==6.2.1
pytest-asyncio==0.14.0
httpx==0.16.1
asgi-lifespan==1.0.1